Test multipart emails, remove URI parsing to path
As noted in the comment for this PR
https://github.com/ianwhite/pickle/pull/17 this strips out URL
params, and it also only tries to open the path where as I think we
should stick with the original intent of opening the exact link from the
email. Your environments/development.rb (or test) should set the proper
base url so that links in emails will go to the test url properly.